Use macro for sandbox files
